Move term config helper into browser
