+++ src/vs/workbench/contrib/terminal/common/terminal.ts
@@ -209,14 +209,6 @@ export interface ITerminalConfiguration {
 
 export const DEFAULT_LOCAL_ECHO_EXCLUDE: ReadonlyArray<string> = ['vim', 'vi', 'nano', 'tmux'];
 
-export interface ITerminalConfigHelper {
-	config: ITerminalConfiguration;
-
-	configFontIsMonospace(): boolean;
-	getFont(w: any): ITerminalFont;
-	showRecommendations(shellLaunchConfig: IShellLaunchConfig): void;
-}
-
 export interface ITerminalFont {
 	fontFamily: string;
 	fontSize: number;
